Extract RealPath from utils:files
Bug: b/542760101
