+++ base/cvd/cuttlefish/common/libs/utils/BUILD.bazel
@@ -120,6 +120,7 @@ cf_cc_library(
         "//cuttlefish/files:file_device_id",
         "//cuttlefish/files:file_exists",
         "//cuttlefish/files:link_or_copy",
+        "//cuttlefish/posix:realpath",
         "//cuttlefish/posix:rename",
         "//cuttlefish/posix:strerror",
         "//cuttlefish/result",
