+++ base/cvd/cuttlefish/common/libs/utils/files.h
@@ -71,8 +71,6 @@ Result<uid_t> FileOwner(const std::string& path);
 // argument.
 std::string AbsolutePath(std::string_view path);
 
-Result<std::string> RealPath(const std::string& path);
-
 std::string CurrentDirectory();
 
 struct FileSizes {
