+++ base/cvd/cuttlefish/host/commands/cvd/cli/BUILD.bazel
@@ -170,6 +170,7 @@ cf_cc_library(
         "//cuttlefish/host/commands/cvd/instances:instance_manager",
         "//cuttlefish/host/commands/cvd/instances/lock",
         "//cuttlefish/host/commands/cvd/utils",
+        "//cuttlefish/posix:realpath",
         "//cuttlefish/posix:strerror",
         "//cuttlefish/posix:symlink",
         "//cuttlefish/result",
