+++ base/cvd/cuttlefish/host/libs/command_util/BUILD.bazel
@@ -42,6 +42,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/config:config_instance_derived",
         "//cuttlefish/host/libs/config:cuttlefish_config",
         "//cuttlefish/posix:readlink",
+        "//cuttlefish/posix:realpath",
         "//cuttlefish/posix:strerror",
         "//cuttlefish/posix:symlink",
         "//cuttlefish/result",
