+++ base/cvd/cuttlefish/host/libs/command_util/snapshot_utils.cc
@@ -36,6 +36,7 @@
 #include "cuttlefish/files/directory_exists.h"
 #include "cuttlefish/files/file_exists.h"
 #include "cuttlefish/posix/readlink.h"
+#include "cuttlefish/posix/realpath.h"
 #include "cuttlefish/posix/symlink.h"
 #include "cuttlefish/result/result.h"
