+++ base/cvd/cuttlefish/host/libs/config/BUILD.bazel
@@ -279,13 +279,11 @@ cf_cc_library(
     name = "fetcher_configs",
     srcs = ["fetcher_configs.cc"],
     hdrs = ["fetcher_configs.h"],
-    depend_on_what_you_use_enabled = False,
     deps = [
-        "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/files:file_exists",
         "//cuttlefish/host/libs/config:fetcher_config",
+        "//cuttlefish/posix:realpath",
         "//cuttlefish/result",
-        "//libbase",
         "@abseil-cpp//absl/log",
         "@abseil-cpp//absl/strings",
         "@abseil-cpp//absl/types:span",
