+++ base/cvd/cuttlefish/host/libs/image_aggregator/BUILD.bazel
@@ -127,9 +127,9 @@ cf_cc_library(
     hdrs = ["sparse_image.h"],
     deps = [
         "//cuttlefish/common/libs/fs",
-        "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/host/libs/config:known_paths",
         "//cuttlefish/host/libs/image_aggregator:disk_image",
+        "//cuttlefish/posix:realpath",
         "//cuttlefish/posix:rename",
         "//cuttlefish/process:execute",
         "//cuttlefish/result",
