+++ base/cvd/cuttlefish/host/libs/image_aggregator/sparse_image.cc
@@ -31,8 +31,8 @@
 #include "sparse/sparse.h"
 
 #include "cuttlefish/common/libs/fs/shared_fd.h"
-#include "cuttlefish/common/libs/utils/files.h"
 #include "cuttlefish/host/libs/config/known_paths.h"
+#include "cuttlefish/posix/realpath.h"
 #include "cuttlefish/posix/rename.h"
 #include "cuttlefish/process/execute.h"
 #include "cuttlefish/result/result.h"
